remove dependency on 'find packages' helpers script
I originally added this because snap fails to install a list of packages
if one of them is already installed, and apt would try to upgrade
packages if they were already installed.

I discovered that snap exits with 0 in that case, so it wouldn't cause
the whole script to fail.

Also discovered the `--no-upgrade` flag to apt, so that behavior is how
I want as well.

So I no longer need the helper script. Leaving it here for now though in
case I come up with a need.